[Writing Tools] Rare WebContent crash in IPC::ArgumentCoder<WebKit::CoreIPCData, void>::encode https://bugs.webkit.org/show_bug.cgi?id=276289 rdar://130595277
Reviewed by Aditya Keerthi.

This crash can happen when attempting to create an NSTextAttachment with a `nil` file wrapper.
Fix by instead just returning an NSAttributedString comprised of a single space in this case (to
maintain the same length as if there were an attachment).

* Source/WebCore/editing/cocoa/HTMLConverter.mm:
(attributedStringWithAttachmentForFileWrapper):
(attributedStringWithAttachmentForElement):

Canonical link: https://commits.webkit.org/280714@main