Drain suspended process message queues more deterministically https://bugs.webkit.org/show_bug.cgi?id=276301 rdar://131274197
Reviewed by Per Arne Vollan.

When we detect that a message queue for a suspended process has filled up, we currently wake it up
for a fixed 3 second duration. This fixed duration might be too short to drain the message queue,
which somewhat defeats the purpose of the logic, or it might be too long to drain the message queue,
giving the process more runtime than we want.

Instead, we now wake up the process for long enough to handle a MainThreadPing, which lets us know
that the process has drained its message queue up to that point. We provide an additional 1 second
of runtime in case the message handlers do any deferred work.

We also want to be able to better debug these types of issues, so I've added some logging which
prints out the most frequent message that is ending up in the stuck message queues.

* Source/WebKit/Platform/IPC/Connection.cpp:
(IPC::Connection::sendMessageImpl):
* Source/WebKit/UIProcess/AuxiliaryProcessProxy.cpp:
(WebKit::AuxiliaryProcessProxy::AuxiliaryProcessProxy):
(WebKit::AuxiliaryProcessProxy::wakeUpTemporarilyForIPC):
* Source/WebKit/UIProcess/AuxiliaryProcessProxy.h:

Canonical link: https://commits.webkit.org/280728@main