Clicking and dragging  video viewer in the page menu drag a layer on top of the other. https://bugs.webkit.org/show_bug.cgi?id=276328 rdar://127512495
Reviewed by NOBODY (OOPS!).

On YouTube.com, clicking and dragging on a video in Video Viewer causes elements on
the page behind the video to come to the forefront. This is because YouTube gives
a draggable=true attribute to the parent div of the video element.

To fix this, this patch sets the media's parent node's draggable attribute to false.
This patch also places the in-window fullscreen specific logic into a separate method
in the MediaController class called _updateControlsForInWindowFullscreenIfNeeded().

* Source/WebCore/Modules/modern-media-controls/media/media-controller.js:
(MediaController.prototype._updateControlsIfNeeded):
(MediaController.prototype._updateControlsForInWindowFullscreenIfNeeded):