[Cocoa] Recognize a second entitlement to allow non-Lockdown Mode processes to launch https://bugs.webkit.org/show_bug.cgi?id=276338 <rdar://127464996>
Reviewed by Per Arne Vollan.

Applications using `WKWebView` that have the `com.apple.developer.web-browser` entitlement can request that a particular page be loaded outside of Lockdown Mode. This allows users to bypass LDM on sites that don't work well under that restricted set of features.

We would like to permit some system processes to launch a plain WKWebView in lockdown mode because of similar compatibility issues. We do not want to use the `web-browser` entitlement, since we do not want the other powerful features that entitlement grants.

We are adding a managed entitlement to support this restricted use case: `com.apple.private.allow-ldm-exempt-webview`

This patch recognizes this new entitlement as an alternative to `com.apple.developer.web-browser` for this specific case.

* Source/WebKit/UIProcess/API/Cocoa/WKWebpagePreferences.mm:
(-[WKWebpagePreferences _setCaptivePortalModeEnabled:]):
(-[WKWebpagePreferences setLockdownModeEnabled:]):

Canonical link: https://commits.webkit.org/280761@main