AX: Iterating over `WebCore::nextVisuallyDistinctCandidate`s can loop infinitely when starting from a role="img" Position https://bugs.webkit.org/show_bug.cgi?id=276383 rdar://131406459
Reviewed by NOBODY (OOPS!).

One example of where we do this is in AXObjectCache:

https://github.com/WebKit/WebKit/blob/b750a5abaa781f9442aa2f3ad68f404362da7bec/Source/WebCore/accessibility/AXObjectCache.cpp#L3336#L3338

In more detail, given this markup:

<div role="img">
    <span style="position:absolute">X</span>
    <span>Y</span>
</div>

When we create a VisiblePosition from this Position:

(Position
  (anchor node: #text 0x159003ec0 length=1 "Y")
  (offset: 0)
  (anchor type: offset in anchor))

We get its `canonicalPosition` in the VisiblePosition constructor, which computes:

(Position
  (anchor node: DIV 0x159003c40)
  (offset: 0)
  (anchor type: before anchor))

This starts iteration for `nextVisuallyDistinctCandidate` back at the beginning of the div, repeating until we get back
to the "Y" position, in turn computing the before-anchor-div position, repeating forever.

This happens because the div is role="img", which was special cased to be `Element::canContainRangeEndPoint()` in:

https://bugs.webkit.org/attachment.cgi?id=229259&action=prettypatch (Find on Page can get stuck in a loop when the search string occurs in an input in a fieldset).

Making it `canContainRangeEndPoint` also makes it `editingIgnoresContent == true`, in turn making it `Position::isCandidate() == true`.

I can no longer reproduce the Find On Page loop bug that this behavior was added to solve, so this patch removes it and
adds a testcase (fast/text/input-fieldset-find-loop.html) confirming there is no loop here. Another testcase
(accessibility/mac/role-img-selection-hang.html) is added to ensure the `nextVisuallyDistinctCandidate` loop is fixed.

Notably, I can also reproduce this hang without any accessibility technology enabled if I make the surrounding body
contenteditable, and add / remove random characters and newlines for a little while (I don't know any exact sequence off-hand).

* LayoutTests/accessibility/mac/role-img-selection-hang-expected.txt: Added.
* LayoutTests/accessibility/mac/role-img-selection-hang.html: Added.
* LayoutTests/fast/text/input-fieldset-find-loop-expected.txt: Added.
* LayoutTests/fast/text/input-fieldset-find-loop.html: Added.
* Source/WebCore/dom/Element.cpp:
(WebCore::Element::canContainRangeEndPoint const):