[iOS] Reconfiguring the device is often switching off the torch rdar://131130489 https://bugs.webkit.org/show_bug.cgi?id=276478
Reviewed by Eric Carlson.

When we reconfigure the device (for frame rate, zoom or size), this often switches off the torch.
To reenable it properly, we asynchronously switch it on again.
We do that asynchronously as switching it on in the same beginConfiguration/commitConfiguration does not seem to work properly.

* Source/WebCore/platform/mediastream/mac/AVVideoCaptureSource.mm:
(WebCore::AVVideoCaptureSource::setFrameRateAndZoomWithPreset):

Canonical link: https://commits.webkit.org/280857@main