[svg] enable support for the 'd' property by default https://bugs.webkit.org/show_bug.cgi?id=276172 rdar://129035240
Reviewed by NOBODY (OOPS!).

Turn on the support for the `d` CSS property by default.

In order to be able to do that, we need to address the performance regression
in the Speedometer3 subtest `React-Stockcharts-SVG` by selectively calling
`setPresentationalHintStyleIsDirty()` under `SVGPathElement::svgAttributeChanged()`
only if we've previously seen a change in the `d` CSS property. In other words,
we only kick in the presentation attribute machinery if we've determined that
the CSS `d` property was used.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/svg/SVGPathElement.cpp:
(WebCore::SVGPathElement::svgAttributeChanged):
(WebCore::SVGPathElement::pathDidChange):
* Source/WebCore/svg/SVGPathElement.h: