REGRESSION (280387@main): 42 media (layout-tests) are constant timeouts https://bugs.webkit.org/show_bug.cgi?id=276492 rdar://131548582
Reviewed by Alexey Proskuryakov.

CFNetwork is failing a strict security policy check when WebKit extensions are bundled in WKTR,
since the bundle ID prefix is not `com.apple` for the bundled extensions. Address this by adding
the NSAppTransportSecurity key in the Info.plist for the WebContent and Networking extensions,
which will allow these tests to succeed.

* Tools/WebKitTestRunner/WebKitTestRunner.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/280904@main