std::remove_cvref definition in wtf/StdLibExtras.h is unnecessary following 280033@main https://bugs.webkit.org/show_bug.cgi?id=276581 rdar://131685071
Reviewed by Yusuke Suzuki.

We bumped the minimum supported GCC version to 11.2.0 in 280033@main, and
it looks like all GTK/WPE EWS bots are deployed with configurations that
comply with this requirement.

As such, the workaround for a missing std::remove_cvref definition in
GCC9< is no longer required. In this patch, we remove both the
aforementioned definition from wtf/StdLibExtras.h, and the CMake
machinery driving its conditional inclusion.

* Source/WTF/wtf/StdLibExtras.h:
* Source/cmake/OptionsCommon.cmake:
* Source/cmake/WebKitCompilerFlags.cmake:

Canonical link: https://commits.webkit.org/280963@main